Added prefabs, mats, and a new script.
Added two new prefabs which are a Coin and Grass Ground prefab. The Grass Ground prefab has a child that is used to show dirt underneath the grass. Materials were added for the grass, dirt, and coin prefabs. Started to work on a script called Player Score in which the players will be able to pick up coins to increase their score.