Added in enemy vertical movement, new win and lose conditions, and ground checks.
Enemies can now move up and down, currently there is a glitch to where they do not all move and they also move at different speeds. Added in a ground check so that the player has to be on ground to jump, basically the player can no longer infinitely jump. Also added in win conditions to where if you die or time runs out but have the correct amount of coins to win you win, and vice versa.